pyproject.toml
[tool.poetry]
name = "floodpipeline"
version = "0.0.1"
authors = [
  "Aklilu Teklesadik <ateklesadik@redcross.nl>",
  "Phuoc Phùng <pphung@redcross.nl>",
  "Jacopo Margutti <jmargutti@redcross.nl>"
]
description = "IBF Pipeline for Floods"
readme = "README.md"
repository = "https://github.com/rodekruis/ibf-pipeline-flood"
packages = [{include = "floodpipeline"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
azure-identity = "*"
azure-keyvault-secrets = "*"
azure-storage-blob = "*"
azure-cosmos = "*"
GeoAlchemy2 = "*"
geopandas = "*"
netCDF4 = "*"
pandas = "*"
pycountry = "*"
python-dotenv = "*"
rasterio = "*"
rasterstats = "*"
SQLAlchemy = "*"
tqdm = "*"
xarray = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"